🔨refactor: Fixtures 이름 변경
