+++ src/test/java/com/example/cloudfour/peopleofdelivery/unit/domain/menu/service/query/MenuQueryServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.example.cloudfour.peopleofdelivery.unit.domain.menu.service.query;
 
-import com.example.cloudfour.peopleofdelivery.mock.TestFixtureFactory;
+import com.example.cloudfour.peopleofdelivery.fixtures.Factory;
 import com.example.cloudfour.peopleofdelivery.domain.menu.dto.MenuResponseDTO;
 import com.example.cloudfour.peopleofdelivery.domain.menu.entity.Menu;
 import com.example.cloudfour.peopleofdelivery.domain.menu.entity.MenuCategory;
@@ -53,7 +53,7 @@ class MenuQueryServiceImplTest {
     @BeforeEach  // 모든 @Test 메소드 실행 전에 매번 호출
     void setUp() {
         // TestFixtureFactory를 활용하여 완전한 테스트 데이터 생성
-        testUserWithAll = TestFixtureFactory.createMockUserWithAll();
+        testUserWithAll = Factory.createMockUserWithAll();
 
         // 연관관계를 통해 필요한 객체들 추출
         testStore = testUserWithAll.getStores().get(0);        // 사용자의 첫 번째 가게
